requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
docutils==0.21.2
flake8==7.1.1
flake8-docstrings==1.7.0
mccabe==0.7.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
snowballstemmer==2.2.0
sqlparse==0.5.3
